# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: MosHosp API
  description: API для платформы помощи нуждающимся
  version: 1.0.0
servers:
  - url: https://api.moshosp.ru
    description: Production server
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        telegram_id:
          type: string
          example: "123456789"
        name:
          type: string
          example: "Иван Петров"
        phone:
          type: string
          example: "+7 (900) 123-45-67"
        email:
          type: string
          example: "ivan@example.com"
        avatar_url:
          type: string
          example: "https://i.pravatar.cc/150?img=1"
        role:
          type: string
          enum: [user, volunteer, admin]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    LoginRequest:
      type: object
      properties:
        telegram_data:
          type: object
          properties:
            id:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            username:
              type: string
            photo_url:
              type: string
            auth_date:
              type: string
            hash:
              type: string
      required:
        - telegram_data
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
...
paths:
  /api/auth/telegram:
    post:
      summary: Авторизация через Telegram
      description: Авторизует пользователя через данные Telegram Login Widget
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Ошибка в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      summary: Обновление токена
      description: Обновляет JWT токен с помощью refresh токена
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Успешное обновление токена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Ошибка в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверный или просроченный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...